requirements.txt
# Core dependencies
python-dotenv==1.0.0
asyncio==3.4.3

# LLM dependencies
groq==0.11.0

# Firecrawl integration
firecrawl-py==0.0.5
langchain-community==0.0.10

# Web framework and WebSocket support
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# CORS and security
fastapi-cors==0.0.6

# Text processing
beautifulsoup4==4.12.2
html2text==2020.1.16
markdownify==0.11.6

# Async HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Logging and monitoring
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
mypy==1.7.1